add integration test for Get Pet

+++ src/main/java/br/com/petland/pet/DataRepository.java
@@ -2,6 +2,6 @@ package br.com.petland.pet;
 
 public interface DataRepository<T> {
 
-	public T getById(int id);
+	public T getById(Long id);
 
 }
